Optional Provide Assembly
Added an optional parameter to Start method, this allows you to pass in the Assembly you'd like to use for version checking. The original was failing when the assembly that needed to be checked was called from a main.exe via a dynamic plug in.
Specifically a Microsoft Office VSTO plug in that I wanted to auto update, when using the original code the GetEntryAssembly() call failed and returned null.