Clear cache and add user-agent on webview2 changelog request
When using webview2, changelog requests can get cached by the webview2 instance and not show the latest changes on the update window. This can usually be handled by proper expiration header usage on the webserver side but I figure that it's a safer bet to always clear  the cache before loading the changelog.
This also makes webview2 use whatever user-agent was set for the HttpUserAgent variable (useful for display logic if you wish to reuse the same changelog on your website and also inside the updater webiew2)